fix: update seatLabelOf function to convert seatNumber to string correctly

+++ contracts/FestivalBlockchainUNQ2026.sol
@@ -95,7 +95,7 @@ contract FestivalBlockchainUNQ2026 is ERC721, ERC2981, Ownable {
         return string.concat(
             string(abi.encodePacked(bytes1(uint8(65 + seat.rowIndex)))),
             "-",
-            seat.seatNumber.toString()
+            uint256(seat.seatNumber).toString()
         );
     }
